[FIX] Added preview and save if Save and publish
isnt shown, cause if we where editing parts we couldnt see the dropdown
nor any button more then the "close" button. So now you atleast see the
save and preview button on parts and not just the close button. Not sure
if this is the best way todo it, but it works at least. (we will see the
same problems if the user dont have full access , cause this code only
fallback onto save and preview)