:bug: Do not Read type from empty meta attribute
It may occur that Enum types have empty meta attributes. This is probably a problem of a serializer which should not create such empty items, but instead not include them at all.
All in all, if empty meta item was read, the entire reader was failing and so did entire request processing resulted in 500 while reading a page.